v0.06 - Defining settings
This time we started our project by creating a configuration file that will be responsible for controlling access to the database and establishing direct contact with the bank. We also created a file to handle the selection of the panel, the user can choose the colors of the panel he wants. we also created some other files that will be useful in the next system updates.